chore: Split code into single session/sessions group.
