+++ ee/api/session_summaries.py
@@ -13,8 +13,8 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
 from posthog.cloud_utils import is_cloud
-from ee.hogai.session_summaries.summary_notebooks import create_summary_notebook
-from ee.hogai.session_summaries.summarize_session import ExtraSummaryContext
+from ee.hogai.session_summaries.summarize_session_group.summary_notebooks import create_summary_notebook
+from ee.hogai.session_summaries.summarize_session.summarize_session import ExtraSummaryContext
 from posthog.api.routing import TeamAndOrgViewSetMixin
 from posthog.clickhouse.query_tagging import tag_queries, Product
 from posthog.models import User
