+++ ee/hogai/session_summaries/summarize_session/stream.py
@@ -1,6 +1,6 @@
 from collections.abc import Generator
 from ee.hogai.utils.asgi import SyncIterableToAsync
-from ee.hogai.session_summaries.summarize_session import ExtraSummaryContext
+from ee.hogai.session_summaries.summarize_session.summarize_session import ExtraSummaryContext
 from posthog.models.team.team import Team
 from posthog.settings import SERVER_GATEWAY_INTERFACE
 from posthog.temporal.ai.session_summary.summarize_session import execute_summarize_session_stream
