+++ ee/hogai/session_summaries/summarize_session_group/patterns.py
@@ -6,8 +6,8 @@ from enum import Enum
 import yaml
 import structlog
 from ee.hogai.session_summaries import SummaryValidationError
-from ee.hogai.session_summaries.output_data import SessionSummarySerializer
-from ee.hogai.session_summaries.summarize_session import SingleSessionSummaryLlmInputs
+from ee.hogai.session_summaries.summarize_session.output_data import SessionSummarySerializer
+from ee.hogai.session_summaries.summarize_session.summarize_session import SingleSessionSummaryLlmInputs
 from ee.hogai.session_summaries.utils import strip_raw_llm_content, unpack_full_event_id
 
 logger = structlog.get_logger(__name__)
