+++ ee/hogai/session_summaries/summarize_session_group/summary_notebooks.py
@@ -4,7 +4,7 @@ from typing import Any
 from posthog.models.notebook.notebook import Notebook
 from posthog.models.user import User
 from posthog.models.team import Team
-from ee.hogai.session_summaries.patterns import (
+from ee.hogai.session_summaries.summarize_session_group.patterns import (
     EnrichedSessionGroupSummaryPatternsList,
     EnrichedSessionGroupSummaryPattern,
     PatternAssignedEventSegmentContext,
