+++ ee/hogai/session_summaries/tests/conftest.py
@@ -6,8 +6,8 @@ from datetime import datetime, UTC
 from openai.types.chat.chat_completion import ChatCompletion, Choice, ChatCompletionMessage
 from ee.hogai.session_summaries.utils import generate_full_event_id
 from posthog.models import Team, User
-from ee.hogai.session_summaries.input_data import COLUMNS_TO_REMOVE_FROM_LLM_CONTEXT
-from ee.hogai.session_summaries.prompt_data import SessionSummaryMetadata, SessionSummaryPromptData
+from ee.hogai.session_summaries.summarize_session.input_data import COLUMNS_TO_REMOVE_FROM_LLM_CONTEXT
+from ee.hogai.session_summaries.summarize_session.prompt_data import SessionSummaryMetadata, SessionSummaryPromptData
 from posthog.schema import SessionBatchEventsQueryResponse, CachedSessionBatchEventsQueryResponse, SessionEventsItem
