+++ ee/hogai/session_summaries/tests/test_input_data.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, MagicMock
 from datetime import datetime
 import json
 
-from ee.hogai.session_summaries.input_data import (
+from ee.hogai.session_summaries.summarize_session.input_data import (
     COLUMNS_TO_REMOVE_FROM_LLM_CONTEXT,
     _skip_event_without_valid_context,
     _get_improved_elements_chain_texts,
@@ -410,8 +410,8 @@ def test_get_paginated_session_events(
     # Prepare mock pages data (add columns to each page)
     processed_pages_data = [(mock_columns, events) if events is not None else (None, None) for events in pages_data]
     with (
-        patch("ee.hogai.session_summaries.input_data.SessionReplayEvents") as mock_replay_events,
-        patch("ee.hogai.session_summaries.input_data.get_team", return_value=mock_team),
+        patch("ee.hogai.session_summaries.summarize_session.input_data.SessionReplayEvents") as mock_replay_events,
+        patch("ee.hogai.session_summaries.summarize_session.input_data.get_team", return_value=mock_team),
     ):
         # Mock the SessionReplayEvents DB model to return different data for each page
         mock_instance = MagicMock()
