+++ ee/hogai/session_summaries/tests/test_prompt_data.py
@@ -1,7 +1,7 @@
 from typing import Any
 import pytest
 from datetime import datetime
-from ee.hogai.session_summaries.prompt_data import SessionSummaryMetadata, SessionSummaryPromptData
+from ee.hogai.session_summaries.summarize_session.prompt_data import SessionSummaryMetadata, SessionSummaryPromptData
 from ee.hogai.session_summaries.utils import prepare_datetime, get_column_index
