+++ posthog/session_recordings/session_recording_api.py
@@ -34,7 +34,7 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.utils.encoders import JSONEncoder
 from ee.hogai.session_summaries.llm.call import get_openai_client
-from ee.hogai.session_summaries.stream import stream_recording_summary
+from ee.hogai.session_summaries.summarize_session.stream import stream_recording_summary
 from posthog.cloud_utils import is_cloud
 import posthog.session_recordings.queries.session_recording_list_from_query
 import posthog.session_recordings.queries.sub_queries.events_subquery
