+++ posthog/temporal/ai/session_summary/activities/patterns.py
@@ -8,7 +8,7 @@ from ee.hogai.session_summaries.llm.consume import (
     get_llm_session_group_patterns_assignment,
     get_llm_session_group_patterns_extraction,
 )
-from ee.hogai.session_summaries.patterns import (
+from ee.hogai.session_summaries.summarize_session_group.patterns import (
     EnrichedSessionGroupSummaryPatternsList,
     RawSessionGroupPatternAssignmentsList,
     RawSessionGroupSummaryPatternsList,
@@ -18,8 +18,11 @@ from ee.hogai.session_summaries.patterns import (
     combine_patterns_with_events_context,
     load_session_summary_from_string,
 )
-from ee.hogai.session_summaries.summarize_session import ExtraSummaryContext, SingleSessionSummaryLlmInputs
-from ee.hogai.session_summaries.summarize_session_group import (
+from ee.hogai.session_summaries.summarize_session.summarize_session import (
+    ExtraSummaryContext,
+    SingleSessionSummaryLlmInputs,
+)
+from ee.hogai.session_summaries.summarize_session_group.summarize_session_group import (
     generate_session_group_patterns_assignment_prompt,
     generate_session_group_patterns_extraction_prompt,
     remove_excessive_content_from_session_summary_for_llm,
