+++ posthog/temporal/ai/session_summary/types/group.py
@@ -1,6 +1,6 @@
 import dataclasses
 
-from ee.hogai.session_summaries.summarize_session import ExtraSummaryContext
+from ee.hogai.session_summaries.summarize_session.summarize_session import ExtraSummaryContext
 from posthog.temporal.ai.session_summary.types.single import SingleSessionSummaryInputs
